Pack200 - new tests and some accompanying bug fixes
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/java/trunk@986626 13f79535-47bb-0310-9956-ffa450edef68

+++ classlib/modules/pack200/src/main/java/org/apache/harmony/pack200/ClassBands.java
@@ -1328,22 +1328,22 @@ public class ClassBands extends BandSet {
             boolean visible, List nameRU, List t, List values, List caseArrayN, List nestTypeRS, List nestNameRU, List nestPairN) {
         if(visible) {
             method_RVPA_bands.addAnnotation(desc, nameRU, t, values, caseArrayN, nestTypeRS, nestNameRU, nestPairN);
-            Integer flag = (Integer) tempMethodFlags.remove(tempMethodFlags.size() - 1);
+            Long flag = (Long) tempMethodFlags.remove(tempMethodFlags.size() - 1);
             if((flag.intValue() & (1<<23)) != 0) {
                 method_RVPA_bands.incrementAnnoN();
             } else {
                 method_RVPA_bands.newEntryInAnnoN();
             }
-            tempMethodFlags.add(new Integer(flag.intValue() | (1<<23)));
+            tempMethodFlags.add(new Long(flag.longValue() | (1<<23)));
         } else {
             method_RIPA_bands.addAnnotation(desc, nameRU, t, values, caseArrayN, nestTypeRS, nestNameRU, nestPairN);
-            Integer flag = (Integer) tempMethodFlags.remove(tempMethodFlags.size() - 1);
-            if((flag.intValue() & (1<<24)) != 0) {
+            Long flag = (Long) tempMethodFlags.remove(tempMethodFlags.size() - 1);
+            if((flag.longValue() & (1<<24)) != 0) {
                 method_RIPA_bands.incrementAnnoN();
             } else {
                 method_RIPA_bands.newEntryInAnnoN();
             }
-            tempMethodFlags.add(new Integer(flag.intValue() | (1<<24)));
+            tempMethodFlags.add(new Long(flag.longValue() | (1<<24)));
         }
     }
